fix: rename odoo mock class
