+++ packages/odoo/test/Adaptor.test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { MockOdooWithSearch } from './MockOdooWithSearch';
+import { MockOdooWithSearch } from './mock/client.js';
 
 import {
   create,
